Correct OpenTK.Mathematics.* namespace for color types and also move individual types out into their own files
We should adhere to well established common conventions when it comes to c# And the general consensus is, one file per type, as well as project folders are namespace providers
