+++ src/OpenTK.Mathematics/Colors/Argb.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenTK.Mathematics.Colors
+{
+    /// <summary>
+    /// Alpha, red, green, blue colorspace.
+    /// </summary>
+    public sealed class Argb : IColorSpace4
+    {
+    }
+}
