+++ src/OpenTK.Mathematics/Colors/Hsl.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenTK.Mathematics.Colors
+{
+    /// <summary>
+    /// Hue, saturation, light colorspace.
+    /// </summary>
+    public sealed class Hsl : IColorSpace3
+    {
+    }
+}
