+++ src/OpenTK.Mathematics/Colors/Hsla.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenTK.Mathematics.Colors
+{
+    /// <summary>
+    /// Hue, saturation, light, alpha colorspace.
+    /// </summary>
+    public sealed class Hsla : IColorSpace4
+    {
+    }
+}
