+++ src/OpenTK.Mathematics/Colors/Hsv.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenTK.Mathematics.Colors
+{
+    /// <summary>
+    /// Hue, saturation, value colorspace.
+    /// </summary>
+    public sealed class Hsv : IColorSpace3
+    {
+    }
+}
