+++ src/OpenTK.Mathematics/Colors/Hsva.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenTK.Mathematics.Colors
+{
+    /// <summary>
+    /// Hue, saturation, value, alpha colorspace.
+    /// </summary>
+    public sealed class Hsva : IColorSpace4
+    {
+    }
+}
