+++ src/OpenTK.Mathematics/Colors/IColorSpace3.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenTK.Mathematics.Colors
+{
+    /// <summary>
+    /// Interface for ColorSpace phantom types with 3 elements.
+    /// </summary>
+    public interface IColorSpace3
+    {
+    }
+}
