+++ src/OpenTK.Mathematics/Colors/IColorSpace4.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenTK.Mathematics.Colors
+{
+    /// <summary>
+    /// Interface for ColorSpace phantom types with 4 elements.
+    /// </summary>
+    public interface IColorSpace4
+    {
+    }
+}
