+++ src/OpenTK.Mathematics/Colors/Rgb.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenTK.Mathematics.Colors
+{
+    /// <summary>
+    /// Red, green, blue colorspace.
+    /// </summary>
+    public sealed class Rgb : IColorSpace3
+    {
+    }
+}
