+++ src/OpenTK.Mathematics/Colors/Rgba.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenTK.Mathematics.Colors
+{
+    /// <summary>
+    /// Red, green, blue, alpha colorspace.
+    /// </summary>
+    public sealed class Rgba : IColorSpace4
+    {
+    }
+}
